migrate getSlide to __slide
