+++ public/components/two-step-signin/two-step-signin.js
@@ -1,11 +1,13 @@
 // @flow
 
-import { EV_DONE } from 'components/two-step-signin/two-step-signin__slide';
+import {
+  EV_DONE,
+  getSlide
+} from 'components/two-step-signin/two-step-signin__slide';
 import { loadComponents } from 'js/load-components';
 import { pageView } from '../analytics/ga';
 
 const selector: string = '.two-step-signin';
-const slideClassName: string = 'two-step-signin__slide';
 
 const ERR_MALFORMED_EVENT: string = 'Something went wrong';
 const ERR_MALFORMED_HTML: string = 'Something went wrong';
@@ -120,22 +122,6 @@ const onSlide = (
   }
 };
 
-const getSlide = ($component: HTMLElement) => {
-  const $slide = $component.querySelector(`.${slideClassName}`);
-  if ($slide) return $slide;
-  throw new Error([
-    ERR_MALFORMED_HTML,
-    $component.querySelector(`.${slideClassName}`)
-  ]);
-};
-
-const getSlideFromFetch = (textHtml: string): HTMLElement => {
-  const $wrapper: HTMLElement = document.createElement('div');
-  $wrapper.innerHTML = textHtml;
-
-  return getSlide($wrapper);
-};
-
 const preservePasswordField = (
   $oldSlide: HTMLElement,
   $newSlide: HTMLElement
@@ -153,7 +139,7 @@ const init = ($component: HTMLElement): void => {
   $component.addEventListener(EV_DONE, (ev: mixed) => {
     if (ev instanceof CustomEvent) {
       const $slide = getSlide($component);
-      const $new = getSlideFromFetch(ev.detail.responseHtml);
+      const $new = ev.detail.$slide;
       const url = ev.detail.url;
 
       if (!$slide || !$new) {
